Initialize system fabric topology table
The system fabric maintains a table of 32 addressable regions. Each
region can be mapped to a particular chip using a 4b topology id. The
index into this table is 5b value (RA[15:19]) also confusingly referred
to as a "topology id".

To try to alleviate confusion in the code, the 4b "chip-unique" topology
id is referred to as `topology_id` and the 5b value referred to in the
RA is called a `topology_idx`.

The topology_id is mapped to the topology_idx as follows:
MODE0 => GGGC --> GGG_0C
MODE1 => GGCC --> 0GG_CC

The topology id table is initialized as follows during p10_sbe_scominit:
topology id table[topology_idx(chip)] = topology_id(chip)

Change-Id: I0b433f53e3c69cc0fdae7168d059beb138f6b5d0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/78202
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: Gregory S. Still <stillgs@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>