Honor DDIMM EFD metadata to choose OMI frequencies
There are two places where the list of supported frequencies are
listed.  First, there are a couple of bytes that list the overall
support.  Second, there are the EFD buckets themselves.  The
previous code walked all the buckets but ignored the summary bytes.
In order to avoid drastically unique SPD images for Gemini, we are
adding support to honor the summary bytes so we can choose specific
speeds to run at.

Change-Id: I82477468eca06d9c23fd8cdc70fa583d49f7d4bf
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/81385
Reviewed-by: ANDRE A MARIN <aamarin@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/83855
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>