Add additional check to verify hostboot istep has completed.
In simics, it's possible that the delays present in p10_do_fw_hb_istep
aren't long enough for hostboot's SPTask thread to respond to an istep
request. This leads to p10_do_fw_hb_istep exiting erroneously thinking
that the istep has completed when it has not. By adding an additional
check when polling for istep completion this should no longer occur and
p10_do_fw_hb_istep should only exit due to a timeout or because the
istep completed.

Change-Id: Ie60a8c82f72162efb4678ec794aacecaf5d4b810
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89771
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: JAYANTH OTHAYOTH <ojayanth@in.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>