Improve TOD Status and Move TOD to TB HWPs
Improve TOD Status and Move TOD to TB HWPs that are
only used for HWP testing purposes.

- Check TFMR bits 25 and 56 error bits in p10_tod_status
- Check TFMR per-thread status bits of TFMR bits 57 through 62
in p10_tod_status
- Return error if any error condition detected in p10_tod_status,
except for TFAC_HOLD_OUT, which can't be cleared after an error
injection.  Needs to be checked outside of p10_tod_status depending
on whether an error is expected or not.
- Don't assume that the Timebase FSM is in the RESET state in
p10_tod_move_tod_to_tb.  Instead, if the FSM is already in
the RUNNING state, then leave in the RUNNING state (new
default behavior).
- Added four new options to p10_tod_move_tod_to_tb via the new HWP
'flags' parameter to transition the Timebase FSM through the different
states depending on whether the Timebase FSM is already in the RUNNING
or not in the RUNNING states:
 * -move_running_reset : If Timebase State Machine is in RUNNING State,
    first transition to RESET State
 * -move_running_send   : If Timebase State Machine is in RUNNING State,
    first transition to SEND_TOD State
 * -move_running_get     : If Timebase State Machine is in RUNNING
    State, first transition to GET_TOD State
 * -move_not_running_reset : If Timebase State Machine is NOT in RUNNING
    State, first transition to RESET State if not already in RESET
- Default behavior is listed above if none of these options are set
  ('flags' = 0)

Change-Id: I2e3264fffa44f325948813f4ccc760863c83483b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105130
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>