Add null pointer check to FFDC gathering functions
Hardware procedure error classes are generated from XML
descriptions. For each error type, these descriptions contain a list
of member pointer variables (contained in instances of ffdc_t) which
need to be set by anyone who creates an error of that type, e.g. with
FAPI_ASSERT.

These member pointers default to NULL. If a programmer forgets to set
all the variables on the error, or if a variable is added to the error
class via the XML and the code that creates the error is not updated
to set the new variable, the code can crash dereferencing the NULL
pointer that the member defaulted to.

This commit adds a check when using those member pointers so that,
instead of crashing on a NULL pointer dereference, the code will log
an error message and indicating that there is a bug in the creation of
the error.

Change-Id: Ic7d57feee09dbc9f5025c6a6afa640bce48913d0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119175
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: Kahn C Evans <kahnevan@us.ibm.com>
Reviewed-by: DAVID A KING <kingboy@us.ibm.com>