Fix max per proc values to reflect P10 values
The values that were set for maxTargetsPerProc were set for P9 which are
incorrect for P10. This results in an erroneous fapi pos calculation in
convertSbeTargInstanceToFapiPos which is called by getFfdcData to lookup
a target based on that fapi position value.

This means that if the target retrieved and return by getFfdcData is
used in a callout it will be the wrong target.

Change-Id: I550707c8b92a6189a79cb81dda1ce819e79405f0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149920
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>