workaround for legacy commit-msg file
The Gerrit server has recently been updated and the commit-msg file from
that server uses newer git command that are not available on older
versions of git. A legacy version of that file was saved and will be
used instead of pulling from the gerrit server if the user has a git
version older that 2.2.0.

Change-Id: I2845b6151f5145873c30a6dbaa4510b359bb4ba1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/161185
Reviewed-by: Doug Gilbert <dgilbert@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>