Add MAX_*_PER_OCMB to convertSbeTargInstanceToFapiPos
There was a bug where the number of PERVASIVE targets on Odyssey OCMBs
was not present in the code which converts SBE target instances to
FAPI_POS.

This commit addresses that issue so the code can convert those values to
be used for callout/deconfig/guard later.

Additionally, adds MAX_DIMM_PER_OCMB to resolve a similar issue.

Change-Id: I320265b5e247f68d44b070081e48c11a262506af
JIRA: PFHB-551
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/164442
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>